Allow fission to create secondary neutrons in fixed source simulations.

+++ src/particle_header.F90
@@ -86,7 +86,7 @@ module particle_header
     logical    :: write_track = .false.
 
     ! Secondary particles created
-    integer    :: n_secondary = 0
+    integer(8) :: n_secondary = 0
     type(Bank) :: secondary_bank(MAX_SECONDARY)
 
   contains
